// src/interfaces.ts
export interface options {
    background?: string;
    size?: Size;
    extraSize?: ExtraSize;
    fontFamily?: FontFamily;
}

interface Size {
    width: number;
    height: number;
}

interface ExtraSize {
    width: number;
    height: number;
}

interface FontFamily {
    url: string,
    fontFamily: string
}